+++ libs/openapi-types/src/v3_0/mod.rs
@@ -1,7 +1,10 @@
+mod component_name;
+pub use component_name::SchemaFieldName;
+
 mod components_object;
 pub use components_object::{
     AllOf, ArrayItems, ComponentsObject, EnumValues, RequiredSchemaFields, SchemaCase,
-    SchemaFieldName, SchemaObject, SchemaProperties, SchemasObject,
+    SchemaObject, SchemaProperties, SchemasObject,
 };
 
 mod format_modifier;
